pyproject.toml
[tool.poetry]
name = "allblue"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
django-ninja = "^0.19.1"
channels = "^4.0.0"
django-environ = "^0.9.0"
daphne = "^4.0.0"
firebase-admin = "^6.0.1"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
